System/Server/manager/requirements.txt
pandas==2.1.3
numpy==1.26.2

mysql-connector-python==8.2.0